# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FDSAP swagger template",
        "title": "Fiber Example API",
        "contact": {
            "name": "FDSAP Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
        "/public/v1/dashboard/dashboardMenu": {
            "post": {
                "description": "Provide the menu to show in kPLUS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DASHBOARD"
                ],
                "summary": "DASHBOARD-MENU",
                "parameters": [
                    {
                        "description": "Insti Code",
                        "name": "getInstiCode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.InstiCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DashboardMenuResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/kplus/get_param": {
            "get": {
                "description": "Provide the data FROM getParam TABLE that will be used by kPLUS upon opening the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kPLUS UPON OPEN"
                ],
                "summary": "API FOR GET PARAM",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.InstiParam"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/kplus/insti_param": {
            "get": {
                "description": "Provide the data FROM insti_param TABLE that will be used by kPLUS upon opening the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kPLUS UPON OPEN"
                ],
                "summary": "API FOR INSTI PARAM",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.InstiParam"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/kplus/kplus_upon_open": {
            "get": {
                "description": "Provide the data that will be used by kPLUS upon opening the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kPLUS UPON OPEN"
                ],
                "summary": "kPLUS UPON OPEN",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetParamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/kplus/splash_screen": {
            "get": {
                "description": "Provide the data FROM c_splash_screen TABLE that will be used by kPLUS upon opening the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kPLUS UPON OPEN"
                ],
                "summary": "API FOR SPLASH SCREEN",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SplashScreenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/transaction/fetch_transaction": {
            "post": {
                "description": "Provides Excel File Report of Transaction in a certain range of date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JANUS REPORT GENERATION"
                ],
                "summary": "API TRANSACTION GENERATION",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/public/v1/transaction/get_path": {
            "get": {
                "description": "Excel File Path Getter To Save into the Database when it was Downloaded",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JANUS REPORT GENERATION"
                ],
                "summary": "API TRANSACTION GENERATION FILE PATH GETTER",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetCellValuePath"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.InstiCodeRequest": {
            "type": "object",
            "properties": {
                "insti_code": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "response.DashboardMenuResponse": {
            "type": "object",
            "properties": {
                "dashboard": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "integer"
                },
                "icon": {
                    "type": "integer"
                },
                "insti_code": {
                    "type": "integer"
                },
                "menu_item": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "soon": {
                    "type": "integer"
                }
            }
        },
        "response.GetCellValuePath": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "response.GetParamResponse": {
            "type": "object",
            "properties": {
                "param_id": {
                    "type": "integer"
                },
                "param_name": {
                    "type": "string"
                },
                "param_value": {
                    "type": "string"
                }
            }
        },
        "response.InstiParam": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "insti_code": {
                    "type": "integer"
                },
                "institution": {
                    "type": "string"
                },
                "telephone_number": {
                    "type": "string"
                }
            }
        },
        "response.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "retCode": {
                    "type": "string"
                }
            }
        },
        "response.SplashScreenResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "last_updated_by": {
                    "type": "integer"
                },
                "last_updated_date": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "redirect_link": {
                    "type": "string"
                },
                "show": {
                    "type": "string"
                },
                "sub_message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "response.TransactionResponse": {
            "type": "object",
            "properties": {
                "got_agent_fee": {
                    "type": "string"
                },
                "got_amount": {
                    "type": "string"
                },
                "got_bank_fee": {
                    "type": "string"
                },
                "got_cid": {
                    "type": "string"
                },
                "got_date": {
                    "type": "string"
                },
                "got_iid": {
                    "type": "string"
                },
                "got_particular": {
                    "type": "string"
                },
                "got_reference": {
                    "type": "string"
                },
                "got_source_account": {
                    "type": "string"
                },
                "got_standing": {
                    "type": "string"
                },
                "got_target": {
                    "type": "string"
                },
                "got_time": {
                    "type": "string"
                },
                "got_total_fee": {
                    "type": "string"
                }
            }
        }
    }
}